Add rough installation instructions / script
This is a very rough scripted install of the DSpace source with LU
customizations. It is intended more of a guide at this point. Under
the correct conditions it can automate the install process, but it
should be made more generic, and not rely on certain things already
being in place.