Need to name the crosswalk oai_etdms instead of etdms
Theses Canada's automated testing interface complains that it cannot
harvest the data in etdms format unless the crosswalk is called
'oai_etdms' instead of just etdms.

Also, this oai_etdms xsl crosswalk is a work in progress that adheres to
most of the Theses Canada requirements. As with the previous java
version that worked prior to Dspace3, it generates metadata, which
it shouldn't. The information that the crosswalk generates is also
LU specific, and would need some work to be generalized.